Rename components, add testing enviroment
