+++ src/__tests__/navbar/Navbar.test.js
@@ -1,0 +1,12 @@
+import { BrowserRouter } from 'react-router-dom';
+import renderer from 'react-test-renderer';
+import Navbar from '../../components/Navbar';
+
+it('renders correctly', () => {
+  const navbar = renderer.create(
+    <BrowserRouter>
+      <Navbar />
+    </BrowserRouter>,
+  ).toJSON();
+  expect(navbar).toMatchSnapshot();
+});
