+++ src/__tests__/stocklist/StockList.test.js
@@ -1,0 +1,12 @@
+import { BrowserRouter } from 'react-router-dom';
+import renderer from 'react-test-renderer';
+import StockList from '../../components/StockList';
+
+it('renders correctly', () => {
+  const navbar = renderer.create(
+    <BrowserRouter>
+      <StockList />
+    </BrowserRouter>,
+  ).toJSON();
+  expect(navbar).toMatchSnapshot();
+});
